requirements.txt
streamlit
pdfplumber
spacy
nltk
google-generativeai
sentence-transformers
python-dotenv
